requirements.txt
Django==1.9.4
django-debug-toolbar==1.4
django-grappelli==2.8.1
django-localflavor==1.2
djangorestframework==3.3.2
extras==0.0.3
fixtures==1.4.0
funcsigs==0.4
fuzzywuzzy==0.8.1
lettuce==0.2.21
linecache2==1.0.0
mock==1.3.0
nose==1.3.7
pbr==1.8.1
Pillow==3.1.1
pyrsistent==0.11.12
python-mimeparse==1.5.1
python-subunit==1.2.0
PyYAML==3.11
six==1.10.0
sure==1.2.24
testtools==2.0.0
traceback2==1.4.0
unittest2==1.1.0

git+https://github.com/ParentJA/example-accounts-django#egg=example-accounts-django